src/v1.5_learning/requirements.txt
# ============================================================
# TryAngle v1.5 MVP Learning Pipeline
# Requirements for Windows RTX 4070 Super
# ============================================================

# ---- Core ----
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0
opencv-python>=4.8.0
tqdm>=4.65.0
pyyaml>=6.0

# ---- Grounding DINO ----
# Install separately: pip install groundingdino-py
# Or clone: https://github.com/IDEA-Research/GroundingDINO
groundingdino-py>=0.1.0

# ---- Depth Anything V2 ----
# Clone from: https://github.com/DepthAnything/Depth-Anything-V2
# Manual installation required
transformers>=4.30.0
huggingface_hub>=0.16.0

# ---- RTMPose (MMPose) ----
# Install: pip install openmim && mim install mmpose mmdet mmcv
openmim>=0.3.9

# ---- Data Processing ----
polars>=0.19.0
pandas>=2.0.0

# ---- Utilities ----
rich>=13.0.0
typer>=0.9.0

# ============================================================
# Installation Order (Windows):
# ============================================================
#
# 1. CUDA 11.8+ 설치 확인
#    nvidia-smi
#
# 2. PyTorch with CUDA
#    pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
#
# 3. 기본 패키지
#    pip install -r requirements.txt
#
# 4. Grounding DINO
#    pip install groundingdino-py
#    (또는 소스 설치: git clone + pip install -e .)
#
# 5. MMPose (RTMPose)
#    pip install openmim
#    mim install mmengine mmcv mmdet mmpose
#
# 6. Depth Anything V2
#    git clone https://github.com/DepthAnything/Depth-Anything-V2
#    pip install -e Depth-Anything-V2
#
# ============================================================
